+++ packages/protocol/lib/compatibility/verify-bytecode.ts
@@ -26,10 +26,13 @@ interface VerificationContext {
   proposal: ProposalTx[]
   Proxy: Truffle.Contract<ProxyInstance>
   web3: Web3
-
-  // TODO: remove this after first smart contracts release
-  isBeforeRelease1: boolean
 }
+interface InitializationData {
+  [contractName: string]: any[]
+}
+
+const ContractNameExtractorRegex = new RegExp(/(.*)Proxy/)
+
 
 // Checks if the given transaction is a repointing of the Proxy for the given
 // contract.
@@ -37,6 +40,9 @@ const isProxyRepointTransaction = (tx: ProposalTx) =>
   tx.contract.endsWith('Proxy') &&
   (tx.function === '_setImplementation' || tx.function === '_setAndInitializeImplementation')
 
+const isProxyRepointAndInitializeTransaction = (tx: ProposalTx) =>
+  tx.contract.endsWith('Proxy') &&
+    tx.function === '_setAndInitializeImplementation'
 const isProxyRepointForIdTransaction = (tx: ProposalTx, contract: string) =>
   tx.contract === `${contract}Proxy` && isProxyRepointTransaction(tx)
 
@@ -81,19 +87,14 @@ const getImplementationAddress = async (contract: string, context: VerificationC
     return getProposedImplementationAddress(contract, context)
   }
 
-  let proxyAddress: string
   if (isLibrary(contract, context)) {
-    proxyAddress = context.libraryAddresses.addresses[contract]
-    // Before the first contracts upgrade libraries are not proxied.
-    if (context.isBeforeRelease1) {
-      return `0x${proxyAddress}`
-    }
-  } else {
-    // contract is registered but we need to check if the proxy is affected by the proposal
-    proxyAddress = isProxyChanged(contract, context)
-      ? getProposedProxyAddress(contract, context)
-      : await context.registry.getAddressForString(contract)
+    return `0x${context.libraryAddresses.addresses[contract]}`
   }
+
+  // contract is registered but we need to check if the proxy is affected by the proposal
+  const proxyAddress = isProxyChanged(contract, context)
+    ? getProposedProxyAddress(contract, context)
+    : await context.registry.getAddressForString(contract)
 
   // at() returns a promise despite Typescript labelling the await as extraneous
   const proxy: ProxyInstance = await context.Proxy.at(
@@ -144,6 +145,54 @@ const dfsStep = async (queue: string[], visited: Set<string>, context: Verificat
   )
 }
 
+const assertValidProposalTransactions = (proposal: ProposalTx[]) => {
+  const invalidTransactions = proposal.filter(
+    (tx) => !isProxyRepointTransaction(tx) && !isRegistryRepointTransaction(tx)
+  )
+  if (invalidTransactions.length > 0) {
+    throw new Error(`Proposal contains invalid release transactions ${invalidTransactions}`)
+  }
+
+  console.info("Proposal contains only valid release transactions!")
+}
+
+const assertValidInitializationData = (
+  artifacts: BuildArtifacts,
+  proposal: ProposalTx[],
+  web3: Web3,
+  initializationData: InitializationData
+) => {
+  const initializingProposals = proposal.filter(isProxyRepointAndInitializeTransaction)
+  const contractsInitialized = new Set()
+  for (const proposalTx of initializingProposals) {
+    const contractName = ContractNameExtractorRegex.exec(proposalTx.contract)[1]
+
+    if (!initializationData[contractName]) {
+      throw new Error(`Initialization Data for ${contractName} could not be found in reference file`)
+    }
+
+    const contract = artifacts.getArtifactByName(contractName)
+    const initializeAbi = contract.abi.find(
+      (abi: any) => abi.type === 'function' && abi.name === 'initialize')
+    const args = initializationData[contractName]
+    const callData = web3.eth.abi.encodeFunctionCall(initializeAbi, args)
+
+    if (callData !== proposalTx.args[1]) {
+      throw new Error(`Intialization Data for ${contractName} in proposal does not match reference file ${initializationData[contractName]}`)
+    }
+
+    contractsInitialized.add(contractName)
+  }
+
+  for (const referenceContractName of Object.keys(initializationData)) {
+    if (!contractsInitialized.has(referenceContractName)) {
+      throw new Error(`Reference file has initialization data for ${referenceContractName}, but proposal does not specify initialization`)
+    }
+  }
+
+  console.info('Initialization Data was verified!')
+}
+
 /*
  * This function will visit all contracts in `contracts` as well as any
  * linked libraries and verify that the compiled and linked source code matches
@@ -156,14 +205,10 @@ export const verifyBytecodes = async (
   proposal: ProposalTx[],
   Proxy: Truffle.Contract<ProxyInstance>,
   web3: Web3,
-  isBeforeRelease1: boolean = false
+  initializationData: InitializationData = {}
 ) => {
-  const invalidTransactions = proposal.filter(
-    (tx) => !isProxyRepointTransaction(tx) && !isRegistryRepointTransaction(tx)
-  )
-  if (invalidTransactions.length > 0) {
-    throw new Error(`Proposal contains invalid release transactions ${invalidTransactions}`)
-  }
+  assertValidProposalTransactions(proposal)
+  assertValidInitializationData(artifacts, proposal, web3, initializationData)
 
   const queue = contracts.filter((contract) => !ignoredContracts.includes(contract))
   const visited: Set<string> = new Set(queue)
@@ -175,7 +220,6 @@ export const verifyBytecodes = async (
     proposal,
     Proxy,
     web3,
-    isBeforeRelease1,
   }
 
   while (queue.length > 0) {