+++ packages/contractkit/CHANGELOG.md
@@ -5,9 +5,9 @@ This package will follow the release process outlined [here](https://docs.celo.o
 
 
 ## Development (not published yet)
-### **[0.4.17--dev]**
+### **[0.4.19--dev]**
 Features
-- [one-line summary] - ( [link PR] )
+- ODIS Client Update - [#5621](https://github.com/celo-org/celo-monorepo/pull/5621)]
 
 Bug Fixes
 - [one-line summary] - ( [link PR] )
@@ -16,6 +16,20 @@ Other Changes
 - [one-line summary - ( [link PR] )
 
 ## Published
+
+### **[0.4.18]** -- 2020-10-30
+Features
+- CIP encryption - [#5091](https://github.com/celo-org/celo-monorepo/pull/5091)
+- ODIS Client Update - [#5621](https://github.com/celo-org/celo-monorepo/pull/5621)
+- KomenciKit - [#5436](https://github.com/celo-org/celo-monorepo/pull/5436)
+
+Other Changes
+- Small change for attestation service - [#5270](https://github.com/celo-org/celo-monorepo/pull/5270)
+- CIP-08 Compressed key support - [#5624](https://github.com/celo-org/celo-monorepo/pull/5624)
+
+### **[0.4.17]** -- 2020-10-27
+Bug Fixes
+- Uses the most up-to-date version of @celo/utils (0.1.21) & fixes backward compatibility issues from the last release
 
 ### **[0.4.16]** -- 2020-10-23
 Features
@@ -33,7 +47,7 @@ Other Changes
 - Filter out MetaTransactionWallet from registered contracts [#5523](https://github.com/celo-org/celo-monorepo/pull/5523)
 - Enable non-singletons in web3/wrapper contract caches [#5518](https://github.com/celo-org/celo-monorepo/pull/5518), [#5507](https://github.com/celo-org/celo-monorepo/pull/5507)
 - `AwsHsmSigner` and `AwsHsmWallet` moved from default exports to named exports - [#5337](https://github.com/celo-org/celo-monorepo/pull/5337)
-- Added `PROXY_SET_IMPLEMENTATION_SIGNATURE` - ( [#5111](https://github.com/celo-org/celo-monorepo/pull/5111) )
+- Added `PROXY_SET_IMPLEMENTATION_SIGNATURE` - [#5111](https://github.com/celo-org/celo-monorepo/pull/5111)
 - Add blockchain parameters to network config - [#5319](https://github.com/celo-org/celo-monorepo/pull/5319)
 - Make initialization of the proxy part of the governance proposal - [#5481](https://github.com/celo-org/celo-monorepo/pull/5481)